refactor(*): move some components into finchers-endpoint
