+++ finchers-core/src/lib.rs
@@ -9,7 +9,6 @@ extern crate mime;
 #[cfg(feature = "from_hyper")]
 extern crate hyper;
 
-mod caller;
 mod never;
 mod string;
 
@@ -18,8 +17,7 @@ pub mod input;
 pub mod output;
 
 pub use bytes::Bytes;
-pub use caller::{BoxedCaller, Caller};
-pub use error::Error;
+pub use error::HttpError;
 pub use input::Input;
 pub use never::Never;
 pub use output::Output;
