+++ finchers-endpoint/src/lib.rs
@@ -12,16 +12,20 @@ pub mod header;
 pub mod method;
 pub mod path;
 
+mod callable;
 mod context;
 mod endpoint;
+mod error;
 
 mod all;
 mod ok;
 mod skip_all;
 
 // re-exports
+pub use callable::Callable;
 pub use context::Context;
 pub use endpoint::{endpoint, Endpoint, IntoEndpoint};
+pub use error::{Error, ErrorKind};
 pub use ext::EndpointExt;
 
 pub use all::{all, All};
